+++ src/symbol/shaping.ts
@@ -14,7 +14,7 @@ import {TextFit} from '../style/style_image';
 import type {ImagePosition} from '../render/image_atlas';
 import {IMAGE_PADDING} from '../render/image_atlas';
 import type {Rect, GlyphPosition} from '../render/glyph_atlas';
-import {type Formatted, type FormattedSection} from '@maplibre/maplibre-gl-style-spec';
+import {type Formatted, type FormattedSection, type VerticalAlign} from '@maplibre/maplibre-gl-style-spec';
 
 enum WritingMode {
     none = 0,
@@ -58,6 +58,18 @@ export type Shaping = {
     verticalizable: boolean;
 };
 
+type ShapingSectionAttributes = {
+    rect: Rect | null;
+    metrics: GlyphMetrics;
+    baselineOffset: number;
+    imageOffset?: number;
+};
+
+type LineShapingSize = {
+    verticalLineContentWidth: number;
+    horizontalLineContentHeight: number;
+};
+
 function isEmpty(positionedLines: Array<PositionedLine>) {
     for (const line of positionedLines) {
         if (line.positionedGlyphs.length !== 0) {
@@ -81,23 +93,28 @@ class SectionOptions {
     fontStack: string;
     // Image options
     imageName: string | null;
+    // Common options
+    verticalAlign: VerticalAlign;
 
     constructor() {
         this.scale = 1.0;
         this.fontStack = '';
         this.imageName = null;
+        this.verticalAlign = 'bottom';
     }
 
-    static forText(scale: number | null, fontStack: string) {
+    static forText(scale: number | null, fontStack: string, verticalAlign: VerticalAlign | null) {
         const textOptions = new SectionOptions();
         textOptions.scale = scale || 1;
         textOptions.fontStack = fontStack;
+        textOptions.verticalAlign = verticalAlign || 'bottom';
         return textOptions;
     }
 
-    static forImage(imageName: string) {
+    static forImage(imageName: string, verticalAlign: VerticalAlign | null) {
         const imageOptions = new SectionOptions();
         imageOptions.imageName = imageName;
+        imageOptions.verticalAlign = verticalAlign || 'bottom';
         return imageOptions;
     }
 
@@ -182,9 +199,28 @@ class TaggedString {
         return this.sectionIndex.reduce((max, index) => Math.max(max, this.sections[index].scale), 0);
     }
 
+    getMaxImageSize(imagePositions: {[_: string]: ImagePosition}): {
+        maxImageWidth: number;
+        maxImageHeight: number;
+    } {
+        let maxImageWidth = 0;
+        let maxImageHeight = 0;
+        for (let i = 0; i < this.length(); i++) {
+            const section = this.getSection(i);
+            if (section.imageName) {
+                const imagePosition = imagePositions[section.imageName];
+                if (!imagePosition) continue;
+                const size = imagePosition.displaySize;
+                maxImageWidth = Math.max(maxImageWidth, size[0]);
+                maxImageHeight = Math.max(maxImageHeight, size[1]);
+            }
+        }
+        return {maxImageWidth, maxImageHeight};
+    }
+
     addTextSection(section: FormattedSection, defaultFontStack: string) {
         this.text += section.text;
-        this.sections.push(SectionOptions.forText(section.scale, section.fontStack || defaultFontStack));
+        this.sections.push(SectionOptions.forText(section.scale, section.fontStack || defaultFontStack, section.verticalAlign));
         const index = this.sections.length - 1;
         for (let i = 0; i < section.text.length; ++i) {
             this.sectionIndex.push(index);
@@ -205,7 +241,7 @@ class TaggedString {
         }
 
         this.text += String.fromCharCode(nextImageSectionCharCode);
-        this.sections.push(SectionOptions.forImage(imageName));
+        this.sections.push(SectionOptions.forImage(imageName, section.verticalAlign));
         this.sectionIndex.push(this.sections.length - 1);
     }
 
@@ -585,6 +621,68 @@ function getAnchorAlignment(anchor: SymbolAnchor) {
     return {horizontalAlign, verticalAlign};
 }
 
+function calculateLineContentSize(
+    imagePositions: {[_: string]: ImagePosition},
+    line: TaggedString,
+    layoutTextSizeFactor: number
+): LineShapingSize {
+    const maxGlyphSize = line.getMaxScale() * ONE_EM;
+    const {maxImageWidth, maxImageHeight} = line.getMaxImageSize(imagePositions);
+
+    const horizontalLineContentHeight = Math.max(maxGlyphSize, maxImageHeight * layoutTextSizeFactor);
+    const verticalLineContentWidth = Math.max(maxGlyphSize, maxImageWidth * layoutTextSizeFactor);
+
+    return {verticalLineContentWidth, horizontalLineContentHeight};
+}
+
+function getVerticalAlignFactor(
+    verticalAlign: VerticalAlign
+) {
+    switch (verticalAlign) {
+        case 'top':
+            return 0;
+        case 'center':
+            return 0.5;
+        default:
+            return 1;
+    }
+}
+
+function getRectAndMetrics(
+    glyphPosition: GlyphPosition,
+    glyphMap: {
+        [_: string]: {
+            [_: number]: StyleGlyph;
+        };
+    },
+    section: SectionOptions,
+    codePoint: number
+): GlyphPosition | null {
+    if (glyphPosition && glyphPosition.rect) {
+        return glyphPosition;
+    }
+
+    const glyphs = glyphMap[section.fontStack];
+    const glyph = glyphs && glyphs[codePoint];
+    if (!glyph) return null;
+
+    const metrics = glyph.metrics;
+    return {rect: null, metrics};
+}
+
+function isLineVertical(
+    writingMode: WritingMode.horizontal | WritingMode.vertical,
+    allowVerticalPlacement: boolean,
+    codePoint: number
+): boolean {
+    return !(writingMode === WritingMode.horizontal ||
+        // Don't verticalize glyphs that have no upright orientation if vertical placement is disabled.
+        (!allowVerticalPlacement && !charHasUprightVerticalOrientation(codePoint)) ||
+        // If vertical placement is enabled, don't verticalize glyphs that
+        // are from complex text layout script, or whitespaces.
+        (allowVerticalPlacement && (whitespace[codePoint] || charInComplexShapingScript(codePoint))));
+}
+
 function shapeLines(shaping: Shaping,
     glyphMap: {
         [_: string]: {
@@ -607,7 +705,7 @@ function shapeLines(shaping: Shaping,
     layoutTextSizeThisZoom: number) {
 
     let x = 0;
-    let y = SHAPING_DEFAULT_OFFSET;
+    let y = 0;
 
     let maxLineLength = 0;
     let maxLineHeight = 0;
@@ -615,17 +713,17 @@ function shapeLines(shaping: Shaping,
     const justify =
         textJustify === 'right' ? 1 :
             textJustify === 'left' ? 0 : 0.5;
+    const layoutTextSizeFactor = ONE_EM / layoutTextSizeThisZoom;
 
     let lineIndex = 0;
     for (const line of lines) {
         line.trim();
 
         const lineMaxScale = line.getMaxScale();
-        const maxLineOffset = (lineMaxScale - 1) * ONE_EM;
         const positionedLine = {positionedGlyphs: [], lineOffset: 0};
         shaping.positionedLines[lineIndex] = positionedLine;
         const positionedGlyphs = positionedLine.positionedGlyphs;
-        let lineOffset = 0.0;
+        let imageOffset = 0.0;
 
         if (!line.length()) {
             y += lineHeight; // Still need a line feed after empty line
@@ -633,78 +731,50 @@ function shapeLines(shaping: Shaping,
             continue;
         }
 
+        const lineShapingSize = calculateLineContentSize(imagePositions, line, layoutTextSizeFactor);
+
         for (let i = 0; i < line.length(); i++) {
             const section = line.getSection(i);
             const sectionIndex = line.getSectionIndex(i);
             const codePoint = line.getCharCode(i);
-            let baselineOffset = 0.0;
-            let metrics = null;
-            let rect = null;
-            let imageName = null;
-            let verticalAdvance = ONE_EM;
-            const vertical = !(writingMode === WritingMode.horizontal ||
-                // Don't verticalize glyphs that have no upright orientation if vertical placement is disabled.
-                (!allowVerticalPlacement && !charHasUprightVerticalOrientation(codePoint)) ||
-                // If vertical placement is enabled, don't verticalize glyphs that
-                // are from complex text layout script, or whitespaces.
-                (allowVerticalPlacement && (whitespace[codePoint] || charInComplexShapingScript(codePoint))));
+            const vertical = isLineVertical(writingMode, allowVerticalPlacement, codePoint);
+
+            let sectionAttributes: ShapingSectionAttributes;
 
             if (!section.imageName) {
-                const positions = glyphPositions[section.fontStack];
-                const glyphPosition = positions && positions[codePoint];
-                if (glyphPosition && glyphPosition.rect) {
-                    rect = glyphPosition.rect;
-                    metrics = glyphPosition.metrics;
-                } else {
-                    const glyphs = glyphMap[section.fontStack];
-                    const glyph = glyphs && glyphs[codePoint];
-                    if (!glyph) continue;
-                    metrics = glyph.metrics;
-                }
-
-                // We don't know the baseline, but since we're laying out
-                // at 24 points, we can calculate how much it will move when
-                // we scale up or down.
-                baselineOffset = (lineMaxScale - section.scale) * ONE_EM;
+                sectionAttributes = shapeTextSection(section, codePoint, vertical, lineShapingSize, glyphMap, glyphPositions);
+                if (!sectionAttributes) continue;
             } else {
-                const imagePosition = imagePositions[section.imageName];
-                if (!imagePosition) continue;
-                imageName = section.imageName;
-                shaping.iconsInText = shaping.iconsInText || true;
-                rect = imagePosition.paddedRect;
-                const size = imagePosition.displaySize;
+                shaping.iconsInText = true;
                 // If needed, allow to set scale factor for an image using
                 // alias "image-scale" that could be alias for "font-scale"
                 // when FormattedSection is an image section.
-                section.scale = section.scale * ONE_EM / layoutTextSizeThisZoom;
+                section.scale = section.scale * layoutTextSizeFactor;
 
-                metrics = {width: size[0],
-                    height: size[1],
-                    left: IMAGE_PADDING,
-                    top: -GLYPH_PBF_BORDER,
-                    advance: vertical ? size[1] : size[0]};
-
-                // Difference between one EM and an image size.
-                // Aligns bottom of an image to a baseline level.
-                const imageOffset = ONE_EM - size[1] * section.scale;
-                baselineOffset = maxLineOffset + imageOffset;
-                verticalAdvance = metrics.advance;
-
-                // Difference between height of an image and one EM at max line scale.
-                // Pushes current line down if an image size is over 1 EM at max line scale.
-                const offset = vertical ? size[0] * section.scale - ONE_EM * lineMaxScale :
-                    size[1] * section.scale - ONE_EM * lineMaxScale;
-                if (offset > 0 && offset > lineOffset) {
-                    lineOffset = offset;
-                }
+                sectionAttributes = shapeImageSection(section, vertical, lineMaxScale, lineShapingSize, imagePositions);
+                if (!sectionAttributes) continue;
+                imageOffset = Math.max(imageOffset, sectionAttributes.imageOffset);
             }
 
+            const {rect, metrics, baselineOffset} = sectionAttributes;
+            positionedGlyphs.push({
+                glyph: codePoint,
+                imageName: section.imageName,
+                x,
+                y: y + baselineOffset + SHAPING_DEFAULT_OFFSET,
+                vertical,
+                scale: section.scale,
+                fontStack: section.fontStack,
+                sectionIndex,
+                metrics,
+                rect
+            });
+
             if (!vertical) {
-                positionedGlyphs.push({glyph: codePoint, imageName, x, y: y + baselineOffset, vertical, scale: section.scale, fontStack: section.fontStack, sectionIndex, metrics, rect});
                 x += metrics.advance * section.scale + spacing;
             } else {
                 shaping.verticalizable = true;
-                positionedGlyphs.push({glyph: codePoint, imageName, x, y: y + baselineOffset, vertical, scale: section.scale, fontStack: section.fontStack, sectionIndex, metrics, rect});
+                const verticalAdvance = section.imageName ? metrics.advance : ONE_EM;
                 x += verticalAdvance * section.scale + spacing;
             }
         }
@@ -713,35 +783,107 @@ function shapeLines(shaping: Shaping,
         if (positionedGlyphs.length !== 0) {
             const lineLength = x - spacing;
             maxLineLength = Math.max(lineLength, maxLineLength);
-            justifyLine(positionedGlyphs, 0, positionedGlyphs.length - 1, justify, lineOffset);
+            justifyLine(positionedGlyphs, 0, positionedGlyphs.length - 1, justify);
         }
 
         x = 0;
-        const currentLineHeight = lineHeight * lineMaxScale + lineOffset;
-        positionedLine.lineOffset = Math.max(lineOffset, maxLineOffset);
+        const maxLineOffset = (lineMaxScale - 1) * ONE_EM;
+        positionedLine.lineOffset = Math.max(imageOffset, maxLineOffset);
+        const currentLineHeight = lineHeight * lineMaxScale + imageOffset;
         y += currentLineHeight;
         maxLineHeight = Math.max(currentLineHeight, maxLineHeight);
         ++lineIndex;
     }
 
     // Calculate the bounding box and justify / align text block.
-    const height = y - SHAPING_DEFAULT_OFFSET;
     const {horizontalAlign, verticalAlign} = getAnchorAlignment(textAnchor);
-    align(shaping.positionedLines, justify, horizontalAlign, verticalAlign, maxLineLength, maxLineHeight, lineHeight, height, lines.length);
+    align(shaping.positionedLines, justify, horizontalAlign, verticalAlign, maxLineLength, maxLineHeight, lineHeight, y, lines.length);
 
-    shaping.top += -verticalAlign * height;
-    shaping.bottom = shaping.top + height;
+    // Calculate the bounding box
+    // shaping.top & shaping.left already include text offset (text-radial-offset or text-offset)
+    shaping.top += -verticalAlign * y;
+    shaping.bottom = shaping.top + y;
     shaping.left += -horizontalAlign * maxLineLength;
     shaping.right = shaping.left + maxLineLength;
+}
+
+function shapeTextSection(
+    section: SectionOptions,
+    codePoint: number,
+    vertical: boolean,
+    lineShapingSize: LineShapingSize,
+    glyphMap: {
+        [_: string]: {
+            [_: number]: StyleGlyph;
+        };
+    },
+    glyphPositions: {
+        [_: string]: {
+            [_: number]: GlyphPosition;
+        };
+    },
+): ShapingSectionAttributes | null {
+    const positions = glyphPositions[section.fontStack];
+    const glyphPosition = positions && positions[codePoint];
+
+    const rectAndMetrics = getRectAndMetrics(glyphPosition, glyphMap, section, codePoint);
+
+    if (rectAndMetrics === null) return null;
+
+    let baselineOffset: number;
+    if (vertical) {
+        baselineOffset = lineShapingSize.verticalLineContentWidth - section.scale * ONE_EM;
+    } else {
+        const verticalAlignFactor = getVerticalAlignFactor(section.verticalAlign);
+        baselineOffset = (lineShapingSize.horizontalLineContentHeight - section.scale * ONE_EM) * verticalAlignFactor;
+    }
+
+    return {
+        rect: rectAndMetrics.rect,
+        metrics: rectAndMetrics.metrics,
+        baselineOffset
+    };
+}
+
+function shapeImageSection(
+    section: SectionOptions,
+    vertical: boolean,
+    lineMaxScale: number,
+    lineShapingSize: LineShapingSize,
+    imagePositions: {[_: string]: ImagePosition},
+): ShapingSectionAttributes | null {
+    const imagePosition = imagePositions[section.imageName];
+    if (!imagePosition) return null;
+    const rect = imagePosition.paddedRect;
+    const size = imagePosition.displaySize;
+
+    const metrics = {width: size[0],
+        height: size[1],
+        left: IMAGE_PADDING,
+        top: -GLYPH_PBF_BORDER,
+        advance: vertical ? size[1] : size[0]};
+
+    let baselineOffset: number;
+    if (vertical) {
+        baselineOffset = lineShapingSize.verticalLineContentWidth - size[1] * section.scale;
+    } else {
+        const verticalAlignFactor = getVerticalAlignFactor(section.verticalAlign);
+        baselineOffset = (lineShapingSize.horizontalLineContentHeight - size[1] * section.scale) * verticalAlignFactor;
+    }
+
+    // Difference between height of an image and one EM at max line scale.
+    // Pushes current line down if an image size is over 1 EM at max line scale.
+    const imageOffset = (vertical ? size[0] : size[1]) * section.scale - ONE_EM * lineMaxScale;
+    
+    return {rect, metrics, baselineOffset, imageOffset};
 }
 
 // justify right = 1, left = 0, center = 0.5
 function justifyLine(positionedGlyphs: Array<PositionedGlyph>,
     start: number,
     end: number,
-    justify: 1 | 0 | 0.5,
-    lineOffset: number) {
-    if (!justify && !lineOffset)
+    justify: 1 | 0 | 0.5) {
+    if (justify === 0)
         return;
 
     const lastPositionedGlyph = positionedGlyphs[end];
@@ -750,10 +892,12 @@ function justifyLine(positionedGlyphs: Array<PositionedGlyph>,
 
     for (let j = start; j <= end; j++) {
         positionedGlyphs[j].x -= lineIndent;
-        positionedGlyphs[j].y += lineOffset;
     }
 }
 
+/**
+ * Aligns the lines based on horizontal and vertical alignment.
+ */
 function align(positionedLines: Array<PositionedLine>,
     justify: number,
     horizontalAlign: number,
@@ -769,7 +913,7 @@ function align(positionedLines: Array<PositionedLine>,
     if (maxLineHeight !== lineHeight) {
         shiftY = -blockHeight * verticalAlign - SHAPING_DEFAULT_OFFSET;
     } else {
-        shiftY = (-verticalAlign * lineCount + 0.5) * lineHeight;
+        shiftY = -verticalAlign * lineCount * lineHeight + 0.5 * lineHeight;
     }
 
     for (const line of positionedLines) {